AX: AXObjectCache should have an Inlines header file rdar://158049733 https://bugs.webkit.org/show_bug.cgi?id=297217
Reviewed by Tyler Wilcock and Ryosuke Niwa.

AXObjectCache is included by dozens of files outside of
accessibility. To make it compile faster, we should put
implementations of inline functions in AXObjectCacheInlines, following
the same pattern used elsewhere in WebCore.

Other classes defined in AXObjectCache.h that aren't actually needed
anywhere outside of accessibility should be moved to their own files.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/SourcesCocoa.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/accessibility/AXAttributeCacheScope.cpp: Added.
(WebCore::AXAttributeCacheScope::AXAttributeCacheScope):
(WebCore::AXAttributeCacheScope::~AXAttributeCacheScope):
* Source/WebCore/accessibility/AXAttributeCacheScope.h: Added.
* Source/WebCore/accessibility/AXLogger.cpp:
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::traverseToOffsetInRange):
(WebCore::composedParentIgnoringDocumentFragments): Deleted.
(WebCore::elementName): Deleted.
(WebCore::hasAccNameAttribute): Deleted.
(WebCore::toSimpleImage): Deleted.
(WebCore::hasRole): Deleted.
(WebCore::hasAnyRole): Deleted.
(WebCore::hasTableRole): Deleted.
(WebCore::hasCellARIARole): Deleted.
(WebCore::hasPresentationRole): Deleted.
(WebCore::isRowGroup): Deleted.
(WebCore::dumpAccessibilityTreeToStderr): Deleted.
(WebCore::AXAttributeCacheEnabler::AXAttributeCacheEnabler): Deleted.
(WebCore::AXAttributeCacheEnabler::~AXAttributeCacheEnabler): Deleted.
* Source/WebCore/accessibility/AXObjectCache.h:
(WebCore::AccessibilityObject::axObjectCache const): Deleted.
(WebCore::AXObjectCache::objectsForIDs const): Deleted.
(WebCore::AXObjectCache::nodeForID const): Deleted.
* Source/WebCore/accessibility/AXObjectCacheInlines.h: Added.
(WebCore::AXObjectCache::objectsForIDs const):
(WebCore::AXObjectCache::nodeForID const):
(WebCore::AXObjectCache::scheduleObjectRegionsUpdate):
(WebCore::AXObjectCache::willUpdateObjectRegions):
(WebCore::AXObjectCache::objectBecameIgnored):
(WebCore::AXObjectCache::objectBecameUnignored):
* Source/WebCore/accessibility/AXTextMarker.cpp:
* Source/WebCore/accessibility/AXTreeStore.h:
(WebCore::AXTreeStore<T>::set): Deleted.
(WebCore::AXTreeStore<T>::add): Deleted.
(WebCore::AXTreeStore<T>::remove): Deleted.
(WebCore::AXTreeStore<T>::contains): Deleted.
(WebCore::AXTreeStore<T>::axObjectCacheForID): Deleted.
(WebCore::AXTreeStore<T>::isolatedTreeForID): Deleted.
(WebCore::AXTreeStore<T>::liveTreeMap): Deleted.
(WebCore::AXTreeStore<T>::isolatedTreeMap): Deleted.
(WebCore::axTreeForID): Deleted.
* Source/WebCore/accessibility/AXTreeStoreInlines.h: Copied from Source/WebCore/accessibility/AXTreeStore.h.
(WebCore::AXTreeStore<T>::set):
(WebCore::AXTreeStore<T>::add):
(WebCore::AXTreeStore<T>::remove):
(WebCore::AXTreeStore<T>::contains):
(WebCore::AXTreeStore<T>::axObjectCacheForID):
(WebCore::AXTreeStore<T>::isolatedTreeForID):
(WebCore::AXTreeStore<T>::liveTreeMap):
(WebCore::AXTreeStore<T>::isolatedTreeMap):
(WebCore::axTreeForID):
* Source/WebCore/accessibility/AXUtilities.cpp: Added.
(WebCore::composedParentIgnoringDocumentFragments):
(WebCore::elementName):
(WebCore::hasAccNameAttribute):
(WebCore::toSimpleImage):
(WebCore::hasRole):
(WebCore::hasAnyRole):
(WebCore::hasTableRole):
(WebCore::hasCellARIARole):
(WebCore::hasPresentationRole):
(WebCore::isRowGroup):
(WebCore::dumpAccessibilityTreeToStderr):
* Source/WebCore/accessibility/AXUtilities.h: Added.
* Source/WebCore/accessibility/AccessibilityList.cpp:
* Source/WebCore/accessibility/AccessibilityMathMLElement.cpp:
* Source/WebCore/accessibility/AccessibilityMediaObject.cpp:
* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::updateChildrenIfNecessary):
* Source/WebCore/accessibility/AccessibilityObject.h:
(WebCore::AccessibilityObject::axObjectCache const):
* Source/WebCore/accessibility/AccessibilityRenderObject.cpp:
* Source/WebCore/accessibility/AccessibilityTable.cpp:
* Source/WebCore/accessibility/AccessibilityTableCell.cpp:
* Source/WebCore/accessibility/AccessibilityTree.cpp:
* Source/WebCore/accessibility/cocoa/AXCoreObjectCocoa.mm:
* Source/WebCore/accessibility/ios/AXRemoteTokenIOS.h: Added.
* Source/WebCore/accessibility/ios/AXRemoteTokenIOS.mm: Added.
(WebCore::Accessibility::newAccessibilityRemoteToken):
* Source/WebCore/accessibility/ios/AccessibilityObjectIOS.mm:
(WebCore::Accessibility::newAccessibilityRemoteToken): Deleted.
* Source/WebCore/accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
(-[WebAccessibilityObjectWrapper accessibilityHitTest:]):
(-[WebAccessibilityObjectWrapper accessibilityContainer]):
* Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp:
* Source/WebCore/accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::generateSubtree):
(WebCore::AXIsolatedTree::updateChildren):
(WebCore::AXIsolatedTree::updateChildrenForObjects):
* Source/WebCore/accessibility/isolatedtree/mac/AXIsolatedObjectMac.mm:
* Source/WebCore/accessibility/mac/AXObjectCacheMac.mm:
(WebCore::AXObjectCache::initializeAXThreadIfNeeded):
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
* Source/WebCore/dom/Document.cpp:
* Source/WebCore/page/Page.cpp:
* Source/WebCore/page/PrintContext.cpp:
* Source/WebCore/rendering/RenderObject.cpp:
* Source/WebKit/UIProcess/ios/WKContentView.mm:
* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:

Canonical link: https://commits.webkit.org/298642@main
