+++ Source/WebCore/SaferCPPExpectations/UncountedLambdaCapturesCheckerExpectations
@@ -23,7 +23,6 @@ Modules/webcodecs/WebCodecsVideoDecoder.cpp
 Modules/webcodecs/WebCodecsVideoEncoder.cpp
 Modules/webdatabase/DatabaseThread.cpp
 Modules/webdatabase/SQLCallbackWrapper.h
-accessibility/AXObjectCache.h
 accessibility/AccessibilityScrollView.cpp
 accessibility/AccessibilityTable.cpp
 accessibility/isolatedtree/AXIsolatedObject.cpp
