+++ Source/WebCore/accessibility/AXTreeStore.h
@@ -64,7 +64,7 @@ class AXTreeStore {
     friend WEBCORE_EXPORT AXTreePtr findAXTree(Function<bool(AXTreePtr)>&&);
 public:
     AXID treeID() const { return m_id; }
-    static WeakPtr<AXObjectCache> axObjectCacheForID(std::optional<AXID>);
+    inline static WeakPtr<AXObjectCache> axObjectCacheForID(std::optional<AXID>);
 #if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
     static RefPtr<AXIsolatedTree> isolatedTreeForID(std::optional<AXID>);
     static void applyPendingChangesForAllIsolatedTrees();
@@ -75,116 +75,20 @@ protected:
         : m_id(axID)
     { }
 
-    static void set(AXID, const AXTreeWeakPtr&);
-    static void add(AXID, const AXTreeWeakPtr&);
-    static void remove(AXID);
-    static bool contains(AXID);
+    inline static void set(AXID, const AXTreeWeakPtr&);
+    inline static void add(AXID, const AXTreeWeakPtr&);
+    inline static void remove(AXID);
+    inline static bool contains(AXID);
 
-    static AXID generateNewID();
+    inline static AXID generateNewID();
     const AXID m_id;
     static Lock s_storeLock;
 private:
-    static HashMap<AXID, WeakPtr<AXObjectCache>>& liveTreeMap();
+    inline static HashMap<AXID, WeakPtr<AXObjectCache>>& liveTreeMap();
 #if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-    static HashMap<AXID, ThreadSafeWeakPtr<AXIsolatedTree>>& isolatedTreeMap() WTF_REQUIRES_LOCK(s_storeLock);
+    inline static HashMap<AXID, ThreadSafeWeakPtr<AXIsolatedTree>>& isolatedTreeMap() WTF_REQUIRES_LOCK(s_storeLock);
 #endif
 };
-
-template<typename T>
-inline void AXTreeStore<T>::set(AXID axID, const AXTreeWeakPtr& tree)
-{
-    ASSERT(isMainThread());
-
-    switchOn(tree,
-        [&] (const WeakPtr<AXObjectCache>& typedTree) {
-            liveTreeMap().set(axID, typedTree);
-        }
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-        , [&] (const ThreadSafeWeakPtr<AXIsolatedTree>& typedTree) {
-            Locker locker { s_storeLock };
-            isolatedTreeMap().set(axID, typedTree.get());
-        }
-#endif
-    );
-}
-
-template<typename T>
-inline void AXTreeStore<T>::add(AXID axID, const AXTreeWeakPtr& tree)
-{
-    ASSERT(isMainThread());
-
-    switchOn(tree,
-        [&] (const WeakPtr<AXObjectCache>& typedTree) {
-            liveTreeMap().add(axID, typedTree);
-        }
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-        , [&] (const ThreadSafeWeakPtr<AXIsolatedTree>& typedTree) {
-            Locker locker { s_storeLock };
-            isolatedTreeMap().add(axID, typedTree.get());
-        }
-#endif
-    );
-}
-
-template<typename T>
-inline void AXTreeStore<T>::remove(AXID axID)
-{
-    if (isMainThread()) {
-        liveTreeMap().remove(axID);
-        return;
-    }
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-    Locker locker { s_storeLock };
-    isolatedTreeMap().remove(axID);
-#endif
-}
-
-template<typename T>
-inline bool AXTreeStore<T>::contains(AXID axID)
-{
-    if (isMainThread())
-        return liveTreeMap().contains(axID);
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-    Locker locker { s_storeLock };
-    return isolatedTreeMap().contains(axID);
-#endif
-}
-
-template<typename T>
-inline WeakPtr<AXObjectCache> AXTreeStore<T>::axObjectCacheForID(std::optional<AXID> axID)
-{
-    return axID ? liveTreeMap().get(*axID) : nullptr;
-}
-
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-template<typename T>
-inline RefPtr<AXIsolatedTree> AXTreeStore<T>::isolatedTreeForID(std::optional<AXID> axID)
-{
-    if (!axID)
-        return nullptr;
-
-    Locker locker { s_storeLock };
-    return isolatedTreeMap().get(*axID).get();
-}
-#endif
-
-template<typename T>
-inline HashMap<AXID, WeakPtr<AXObjectCache>>& AXTreeStore<T>::liveTreeMap()
-{
-    ASSERT(isMainThread());
-
-    static NeverDestroyed<HashMap<AXID, WeakPtr<AXObjectCache>>> map;
-    return map;
-}
-
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-template<typename T>
-inline HashMap<AXID, ThreadSafeWeakPtr<AXIsolatedTree>>& AXTreeStore<T>::isolatedTreeMap()
-{
-    static NeverDestroyed<HashMap<AXID, ThreadSafeWeakPtr<AXIsolatedTree>>> map;
-    return map;
-}
-#endif
 
 template<typename T>
 inline AXID AXTreeStore<T>::generateNewID()
@@ -196,18 +100,6 @@ inline AXID AXTreeStore<T>::generateNewID()
         axID = AXID::generate();
     } while (liveTreeMap().contains(*axID));
     return *axID;
-}
-
-template<typename T>
-Lock AXTreeStore<T>::s_storeLock;
-
-inline AXTreePtr axTreeForID(std::optional<AXID> axID)
-{
-#if ENABLE(ACCESSIBILITY_ISOLATED_TREE)
-    if (!isMainThread())
-        return AXTreeStore<AXIsolatedTree>::isolatedTreeForID(axID);
-#endif
-    return AXTreeStore<AXObjectCache>::axObjectCacheForID(axID);
 }
 
 } // namespace WebCore
