+++ Source/WebCore/accessibility/isolatedtree/AXIsolatedObject.cpp
@@ -32,6 +32,8 @@
 #include "AXIsolatedTree.h"
 #include "AXLogger.h"
 #include "AXLoggerBase.h"
+#include "AXObjectCache.h"
+#include "AXObjectCacheInlines.h"
 #include "AXSearchManager.h"
 #include "AXTextMarker.h"
 #include "AXTextRun.h"
