+++ Source/WebCore/page/PrintContext.cpp
@@ -21,6 +21,9 @@
 #include "config.h"
 #include "PrintContext.h"
 
+#include "AXIsolatedTree.h"
+#include "AXObjectCache.h"
+#include "AXObjectCacheInlines.h"
 #include "CommonAtomStrings.h"
 #include "ContainerNodeInlines.h"
 #include "ElementTraversal.h"
